requirements.txt
aiohttp==3.10.10
aiosignal==1.3.1
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
pip==24.2
py-cord==2.6.1
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
